[REFACTOR] #109 프래그먼트 화면 이동 로직 개선

+++ app/src/main/java/com/example/banchan/presentation/adapter/orderlist/OrderListAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.ListAdapter
 import com.example.banchan.domain.model.OrderListModel
 
 class OrderListAdapter(
-    private val onOrderClickListener: (() -> Unit)
+    private val onOrderClickListener: ((Long) -> Unit)
 ): ListAdapter<OrderListModel, OrderListViewHolder>(diffUtil) {
 
     companion object {
